rename(page_buffer_util): support -> util
